requirements.txt
langchain==0.3.20
langchain-openai==0.3.8
nx-arangodb==1.3.0
python-arango==8.1.6
langgraph==0.3.5
pydantic==2.10.6
python-dotenv==1.0.1
tqdm==4.67.1
networkx==3.4
matplotlib==3.10.1
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
cudf_cu12==25.2.2
cugraph_cu12==25.2.0
nx_cugraph_cu12==25.2.0
scipy==1.15.2
langchain_community==0.3.19
streamlit==1.43.1